requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg[binary]==3.2.12

# Authentication
djangorestframework-simplejwt==5.3.1

# Password Hashing
argon2-cffi==23.1.0
bcrypt==4.1.2

# API Documentation
drf-spectacular==0.27.1

# Environment Variables
python-dotenv==1.0.1

# Production Server
gunicorn==21.2.0

# Utilities
pytz==2024.1

# Reportes y Exportación
reportlab==4.0.7
openpyxl==3.1.2
